Determine if token is iOS or android, #37
The current implementation is rather hacky. Ideally the platformType should determine what type of token it is however there has been some problem within the ionic app to retrieve the proper platform data. This is needed because we need to know which platform endpoint the device needs to be added to.